chore: rename riengine ric
