+++ pkg/ric/doc.go
@@ -1,0 +1,2 @@
+// Package ric contains the richer-input compiler.
+package ric
