--- pkg/riengine/doc.go
@@ -1,2 +1,0 @@
-// Package riengine contains the richer input engine.
-package riengine
